+++ client-rpc/server/src/main.rs
@@ -1,5 +1,4 @@
 mod program;
-mod rpc;
 mod server;
 
 fn main() {