+++ client-rpc/src/lib.rs
@@ -1,8 +1,21 @@
-mod program;
-/// Used by c api
-pub mod rpc;
-mod server;
+use std::fmt::Debug;
 
-pub fn run() {
-    crate::program::run_electron();
+pub mod rpc;
+
+pub fn to_rpc_error<E: ToString + Debug>(error: E) -> jsonrpc_core::Error {
+    log::error!("{:?}", error);
+    jsonrpc_core::Error {
+        code: jsonrpc_core::ErrorCode::InternalError,
+        message: error.to_string(),
+        data: None,
+    }
+}
+
+pub fn rpc_error_from_string(error: String) -> jsonrpc_core::Error {
+    log::error!("{}", error);
+    jsonrpc_core::Error {
+        code: jsonrpc_core::ErrorCode::InternalError,
+        message: error,
+        data: None,
+    }
 }